test: rename test files to match pytest's expected format
